Move API logs to lib/api-logs and rename capture-request-log
